fix(refresh-token-flow): add a layer of indirection to ensure refresh flow always calls the latest refresh tokens
Previously, the refresh token flow was coupled to the React render lifecycle as the AuthContext only gets updated
when its useLayoutEffect detects the change in React useState. This
means that multiple refresh token flows in a row could happen before the useLayoutEffect is triggered, leading
to subsequent refresh flows being called with the outdated refresh token inside the axios response interceptor.